+++ plume/src/main/kotlin/io/github/plume/oss/Extractor.kt
@@ -25,7 +25,7 @@ import io.github.plume.oss.drivers.OverflowDbDriver
 import io.github.plume.oss.metrics.ExtractorTimeKey
 import io.github.plume.oss.metrics.PlumeTimer
 import io.github.plume.oss.options.ExtractorOptions
-import io.github.plume.oss.passes.SCPGPass
+import io.github.plume.oss.passes.DataFlowPass
 import io.github.plume.oss.passes.graph.BaseCPGPass
 import io.github.plume.oss.passes.graph.CGPass
 import io.github.plume.oss.passes.method.MethodStubPass
@@ -40,6 +40,7 @@ import io.github.plume.oss.util.ResourceCompilationUtil
 import io.github.plume.oss.util.ResourceCompilationUtil.COMP_DIR
 import io.github.plume.oss.util.ResourceCompilationUtil.TEMP_DIR
 import io.github.plume.oss.util.SootToPlumeUtil
+import io.shiftleft.codepropertygraph.generated.EdgeTypes.INHERITS_FROM
 import io.shiftleft.codepropertygraph.generated.NodeKeyNames.*
 import io.shiftleft.codepropertygraph.generated.NodeTypes.*
 import io.shiftleft.codepropertygraph.generated.nodes.NewFileBuilder
@@ -79,7 +80,6 @@ class Extractor(val driver: IDriver) {
     init {
         File(COMP_DIR).let { f -> if (f.exists()) f.deleteRecursively(); f.deleteOnExit() }
         checkDriverConnection(driver)
-        PlumeTimer.reset()
     }
 
     /**
@@ -169,7 +169,7 @@ class Extractor(val driver: IDriver) {
          */
         logger.info("Building internal program structure and type information")
         val csToBuild = mutableListOf<SootClass>()
-        PlumeTimer.measure(ExtractorTimeKey.BASE_CPG_BUILDING) {
+        PlumeTimer.measure(ExtractorTimeKey.PROGRAM_STRUCTURE_BUILDING) {
             // First read the existing TYPE, TYPE_DECL, and FILEs from the driver and load it into the cache
             populateGlobalTypeCache()
             pipeline(
@@ -189,7 +189,7 @@ class Extractor(val driver: IDriver) {
             Obtain all referenced types from fields, returns, and locals
          */
         val ts = mutableListOf<Type>()
-        PlumeTimer.measure(ExtractorTimeKey.BASE_CPG_BUILDING) {
+        PlumeTimer.measure(ExtractorTimeKey.PROGRAM_STRUCTURE_BUILDING) {
             val fieldsAndRets = csToBuild.map { c -> c.fields.map { it.type } + c.methods.map { it.returnType } }
                 .flatten().toSet()
             val locals = sootUnitGraphs.map { it.body.locals + it.body.parameterLocals }
@@ -202,13 +202,13 @@ class Extractor(val driver: IDriver) {
          */
         logger.info("Building primitive type information")
         logger.debug("All referenced types: ${ts.groupBy { it.javaClass }.mapValues { it.value.size }}}")
-        PlumeTimer.measure(ExtractorTimeKey.BASE_CPG_BUILDING) {
+        PlumeTimer.measure(ExtractorTimeKey.PROGRAM_STRUCTURE_BUILDING) {
             pipeline(
                 GlobalTypePass(driver)::runPass
             ).invoke(ts)
         }
         /*
-            TODO: Handle inheritance
+            Obtain inheritance information
         */
         logger.info("Obtaining class hierarchy")
         val parentToChildCs: MutableList<Pair<SootClass, Set<SootClass>>> = mutableListOf()
@@ -231,6 +231,19 @@ class Extractor(val driver: IDriver) {
                 FileAndPackagePass(driver)::runPass,
                 ExternalTypePass(driver)::runPass,
             ).invoke(filteredExtTypes.map { it.sootClass })
+        }
+        /*
+            Build inheritance edges, i.e.
+            TYPE_DECL -INHERITS_FROM-> TYPE
+        */
+        PlumeTimer.measure(ExtractorTimeKey.PROGRAM_STRUCTURE_BUILDING) {
+            parentToChildCs.forEach { (c, children) ->
+                GlobalCache.getType(c.type.toQuotedString())?.let { t ->
+                    children.intersect(csToBuild)
+                        .mapNotNull { child -> GlobalCache.getTypeDecl(child.type.toQuotedString()) }
+                        .forEach { td -> driver.addEdge(td, t, INHERITS_FROM) }
+                }
+            }
         }
         csToBuild.clear() // Done using csToBuild
         /*
@@ -255,9 +268,11 @@ class Extractor(val driver: IDriver) {
                 .coerceAtLeast(1)
                 .coerceAtMost(Runtime.getRuntime().availableProcessors())
             val channel = Channel<DeltaGraph>()
-            logger.info("""
+            logger.info(
+                """
                 Building ${headsToBuild.size} method heads and ${bodiesToBuild.size} method bodies over $nThreads thread(s).
-            """.trimIndent())
+            """.trimIndent()
+            )
             // Create jobs in chunks and submit these jobs to a thread pool
             val threadPool = Executors.newFixedThreadPool(nThreads)
             try {
@@ -300,8 +315,8 @@ class Extractor(val driver: IDriver) {
         /*
             Method body level analysis - only done on new/updated methods
          */
-        logger.info("Running SCPG passes")
-        PlumeTimer.measure(ExtractorTimeKey.SCPG_PASSES) { SCPGPass(driver).runPass() }
+        logger.info("Running data flow passes")
+        PlumeTimer.measure(ExtractorTimeKey.DATA_FLOW_PASS) { DataFlowPass(driver).runPass() }
         GlobalCache.methodBodies.clear()
         return this
     }