+++ CHANGELOG.md
@@ -12,10 +12,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `CONTAINS` edges are generated for `METHOD` to body vertices.
 - `ListMapper` to process Scala lists to a serialized string and back. More formally processing Scala lists to and from
   OverflowDB node objects.
+- Handle inheritance edges i.e. `TYPE_DECL -INHERITS_FROM-> TYPE`
 
 ### Changed
 
 - `BaseCpgPass` now uses a local cache for method body nodes instead of relying solely on `GlobalCache`
+- `SCPGPass` now known as `DataFlowPass` as all passes now come from `dataflowengineoss`.
+- Added `PROGRAM_STRUCTURE` to timer keys.
 
 ## [0.3.5] - 2021-03-17